made OutputBuilder disposable moved Parsing to static file (for ease of edits)
