--- systemdata/Files/SplashScreens/wm0.txt
@@ -1,10 +1,0 @@
-
-<%green%>
- _      ____           ____  _____  _____             __
-| | /| / / /  ___ ___ / /  |/  / / / / _ \  ___  ___ / /_
-| |/ |/ / _ \/ -_) -_) / /|_/ / /_/ / // / / _ \/ -_) __/
-|__/|__/_//_/\__/\__/_/_/  /_/\____/____(_)_//_/\__/\__/
-<%n%>
-{3}
-
-Welcome to {0} {1}.
