--- systemdata/Files/SplashScreens/wm10.txt
@@ -1,28 +1,0 @@
-
-<%green%>
-       @@@@@@      @@@@@@@@
-      @@    @@@  @@@      @@
-      @        @@@@         @    @   @   @  @                    @
-     @@         @@          @@   @  @ @  @  @                    @
-    @@          @@           @   @  @ @  @  @ @@    @@@    @@@   @
-    @@          @@           @   @  @ @  @  @@  @  @   @  @   @  @
-    @@         @@@@@         @   @ @   @ @  @   @  @   @  @   @  @
-    @@      @@@@@@@@@@@     @@   @ @   @ @  @   @  @@@@@  @@@@@  @
-    @@     @@@@@@@@@@@@     @@   @ @   @ @  @   @  @      @      @
-    @@@    @@@@@@@@@@@@@   @@     @     @   @   @  @   @  @   @  @
-     @@   @@@@@@@@@@@@@@@ @@      @     @   @   @   @@@    @@@   @
-      @@@ @@@ @  @  @ @@@@@
-        @@@@@  @ @ @  @@@@             @     @   @     @  @@@@@
-         @@@    @@@    @@@             @@   @@   @     @  @    @
-         @@@@ @ @@@ @ @@@@             @@   @@   @     @  @     @
-         @@@    @@@    @@@             @ @ @ @   @     @  @     @
-         @@@   @ @ @   @@@             @ @ @ @   @     @  @     @
-          @@@ @  @  @ @@@              @ @ @ @   @     @  @     @
-          @@@@   @   @@@@              @ @ @ @   @     @  @     @
-           @@@@@@@@@@@@@@              @  @  @    @   @   @    @
-            @@@@@@@@@@@                @  @  @     @@@    @@@@@
-              @@@@@@
-<%n%>
-{3}
-
-Welcome to {0} {1}.
