--- systemdata/Files/SplashScreens/wm2.txt
@@ -1,12 +1,0 @@
-
-<%green%>
- _       _  _                    __    __   __   _     _  _____  
-(_)  _  (_)(_)      ____   ____ (__)  (__)_(__) (_)   (_)(_____) 
-(_) (_) (_)(_)__   (____) (____) (_) (_) (_) (_)(_)   (_)(_)  (_)
-(_) (_) (_)(____) (_)_(_)(_)_(_) (_) (_) (_) (_)(_)   (_)(_)  (_)
-(_)_(_)_(_)(_) (_)(__)__ (__)__  (_) (_)     (_)(_)___(_)(_)__(_)
- (__) (__) (_) (_) (____) (____)(___)(_)     (_) (_____) (_____) 
-<%n%>
-{3}
-
-Welcome to {0} {1}.
