--- systemdata/Files/SplashScreens/wm3.txt
@@ -1,11 +1,0 @@
-
-<%green%>
-(_)(_)(_)| |                  | | (_______)(_)   (_)(______) 
- _  _  _ | |__   _____  _____ | |  _  _  _  _     _  _     _ 
-| || || ||  _ \ | ___ || ___ || | | ||_|| || |   | || |   | |
-| || || || | | || ____|| ____|| | | |   | || |___| || |__/ / 
- \_____/ |_| |_||_____)|_____) \_)|_|   |_| \_____/ |_____/  
-<%n%>
-{3}
-
-Welcome to {0} {1}.
