--- systemdata/Files/SplashScreens/wm4.txt
@@ -1,11 +1,0 @@
-
-<%green%>
-Y8b Y8b Y888P 888                     888     e   e     8888 8888 888 88e  
- Y8b Y8b Y8P  888 ee   ,e e,   ,e e,  888    d8b d8b    8888 8888 888 888b 
-  Y8b Y8b Y   888 88b d88 88b d88 88b 888   e Y8b Y8b   8888 8888 888 8888D
-   Y8b Y8b    888 888 888   , 888   , 888  d8b Y8b Y8b  8888 8888 888 888P 
-    Y8P Y     888 888  "YeeP"  "YeeP" 888 d888b Y8b Y8b 'Y88 88P' 888 88"  
-<%n%>
-{3}
-
-Welcome to {0} {1}.
