--- systemdata/Files/SplashScreens/wm5.txt
@@ -1,12 +1,0 @@
-
-<%green%>
- _  _  _  _                    _  ______   _     _  _____    
-| || || || |                  | ||  ___ \ | |   | |(____ \   
-| || || || | _    ____   ____ | || | _ | || |   | | _   \ \  
-| ||_|| || || \  / _  ) / _  )| || || || || |   | || |   | | 
-| |___| || | | |( (/ / ( (/ / | || || || || |___| || |__/ /  
- \______||_| |_| \____) \____)|_||_||_||_| \______||_____/   
-<%n%>
-{3}
-
-Welcome to {0} {1}.
