--- systemdata/Files/SplashScreens/wm6.txt
@@ -1,10 +1,0 @@
-
-<%green%>
-/ / /\ \ \| |__    ___   ___ | |  /\/\   /\ /\    /   \
-\ \/  \/ /| '_ \  / _ \ / _ \| | /    \ / / \ \  / /\ /
- \  /\  / | | | ||  __/|  __/| |/ /\/\ \\ \_/ / / /_// 
-  \/  \/  |_| |_| \___| \___||_|\/    \/ \___/ /___,'  
-<%n%>
-{3}
-
-Welcome to {0} {1}.
