--- systemdata/Files/SplashScreens/wm7.txt
@@ -1,11 +1,0 @@
-
-<%green%>
-__        __  _                     _   __  __   _   _   ____  
-\ \      / / | |__     ___    ___  | | |  \/  | | | | | |  _ \ 
- \ \ /\ / /  | '_ \   / _ \  / _ \ | | | |\/| | | | | | | | | |
-  \ V  V /   | | | | |  __/ |  __/ | | | |  | | | |_| | | |_| |
-   \_/\_/    |_| |_|  \___|  \___| |_| |_|  |_|  \___/  |____/ 
-<%n%>
-{3}
-
-Welcome to {0} {1}.
