--- systemdata/Files/SplashScreens/wm8.txt
@@ -1,11 +1,0 @@
-
-<%green%>
-| |  | || |                | ||  \/  || | | ||  _  \
-| |  | || |__    ___   ___ | || .  . || | | || | | |
-| |/\| || '_ \  / _ \ / _ \| || |\/| || | | || | | |
-\  /\  /| | | ||  __/|  __/| || |  | || |_| || |/ / 
- \/  \/ |_| |_| \___| \___||_|\_|  |_/ \___/ |___/  
-<%n%>
-{3}
-
-Welcome to {0} {1}.
