--- systemdata/Files/SplashScreens/wm9.txt
@@ -1,11 +1,0 @@
-
-<%green%>
-/  \    /  \|  |__    ____    ____  |  |    /     \  |    |   \\______ \   
-\   \/\/   /|  |  \ _/ __ \ _/ __ \ |  |   /  \ /  \ |    |   / |    |  \  
- \        / |   Y  \\  ___/ \  ___/ |  |__/    Y    \|    |  /  |    `   \ 
-  \__/\  /  |___|  / \___  > \___  >|____/\____|__  /|______/  /_______  / 
-       \/        \/      \/      \/               \/                   \/  
-<%n%>
-{3}
-
-Welcome to {0} {1}.
